pyproject.toml
[tool.poetry]
name = "stronakola"
version = "0.1.0"
description = ""
authors = ["Ciastelix <tenlegitmati@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
asyncio = "^3.4.3"
python-dotenv = "^0.21.1"
speechrecognition = "^3.9.0"
pyaudio = "^0.2.13"
pyttsx3 = "^2.90"
setuptools-rust = ">=0.12,<0.13"
pytest = ">=5.4,<6.0"
pytest-asyncio = ">=0.15,<1.0"
pytest-dotenv = ">=0.5,<1.0"
maturin = "^0.14.12"
discord-py = {version = "1.7.3", extras = ["voice"]}
discord = {version = "1.7.3", extras = ["voice"]}
fastapi-utils = "^0.2.1"
dependency-injector = "^4.41.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1a1", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
